+++ packages/flutter_spanned_controller/CHANGELOG.md
@@ -7,7 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- `ToggleStateListener`, which is a generalization of `AttributeListener`
+- `ToggleStateListener`, which is a generalization of `AttributeListener`.
+- `StringDiff.isEmpty` and `StringDiff.isNotEmpty`.
 
 ### Fixed